Move results to profile reducer
